chore: Refactor TodoApp component and add login functionality
This commit refactors the TodoApp component in the src/components/todo/TodoApp.jsx file. It adds a login functionality to the application by introducing a LoginComponent and a WelcomeComponent. The LoginComponent allows users to enter their username and password, and upon successful authentication, displays a success message. If the authentication fails, an error message is shown. This change enhances the user experience and improves the overall functionality of the application.